backend/requirements.txt
Flask
Flask-Cors
pydub==0.25.1
openai-whisper
numpy
flask-socketio
python-dotenv
google-generativeai
PyMuPDF